pyproject.toml
[tool.poetry]
name = "openweather_task"
version = "0.1.0"
description = ""
authors = ["Anton Petrov <anton.petrov.swe@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.61.1"
databases = {extras = ["postgresql"], version = "^0.4.0"}
FastAPI-SQLAlchemy = "^0.2.1"
pydantic = "^1.6.1"
alembic = "^1.4.3"
gunicorn = "^20.0.4"
uvloop = "^0.14.0"
uvicorn = {extras = ["standard"], version = "^0.12.2"}

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
black = "^20.8b1"
mypy = "^0.790"
isort = "^5.6.4"
flake8 = "^3.8.4"
pytest-cov = "^2.10.1"
pre-commit = "^2.7.1"
pytest-asyncio = "^0.14.0"
docker = "^4.3.1"
async-asgi-testclient = "^1.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
